% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_a_optimality.R
\name{calc.a.optimality}
\alias{calc.a.optimality}
\title{calculate A-optimality}
\usage{
calc.a.optimality(infos, prior = NULL, summed = TRUE)
}
\arguments{
\item{infos}{block information as returned from calc.block.infos}

\item{prior}{matrix of prior covariances, if given, posterior information is computed, defaults to NULL}

\item{summed}{logical, should the information be summed across blocks? defaults to TRUE}
}
\value{
numeric vector of A-optimality for each person
}
\description{
calculate A-optimality
}
