% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_info_block.R
\name{calc.info.block}
\alias{calc.info.block}
\title{MFC block information}
\usage{
calc.info.block(
  FUN,
  traits = NULL,
  int,
  loads,
  uni,
  K,
  nb,
  which.blocks = NULL,
  responses = NULL,
  ...
)
}
\arguments{
\item{FUN}{function to compute response probability}

\item{traits}{matrix of traits, rows = persons, columns = traits}

\item{int}{vector of item intercepts}

\item{loads}{matrix of item loadings, rows = items, columns = traits}

\item{uni}{matrix of item uniquenesses}

\item{K}{numeric, number of blocks}

\item{nb}{numeric, block size}

\item{which.blocks}{vector, optional, indices of blocks to calculate info for, defaults to 1:K}

\item{responses}{matrix of indices of observed rank orders, rows = persons, cols = blocks, defaults to NULL}

\item{...}{other arguments passed to FUN}
}
\value{
list of length N, with entries (block,trait,trait) with Fisher (= expected) information; 
if observed information is required: list with two elements: expected and observed information, 
each is a list of length N, with entries (block,trait,trait)
}
\description{
across persons and items
}
