% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_info_plot.R
\name{calc.info.plot}
\alias{calc.info.plot}
\title{compute information for plots}
\usage{
calc.info.plot(
  tr.levels,
  fix.levels = NULL,
  fix.level.others = 0,
  K,
  loads,
  which.blocks,
  ...
)
}
\arguments{
\item{tr.levels}{vector, trait levels that vary}

\item{fix.levels}{vector, trait levels that are fixed, only needed for block size > 2, therefore defaults to NULL}

\item{fix.level.others}{numeric, level for traits not measured by current block (if traits is not given), defaults to 0}

\item{K}{number of blocks}

\item{loads}{matrix of item loadings}

\item{which.blocks}{blocks to calculate information for, defaults to 1:K}

\item{...}{FUN to calculate information and further parameters passed to FUN}
}
\value{
information for relevant trait combinations for all blocks, and for which trait pairs (combinations) information is
}
\description{
computes information needed for plot.block()
}
\details{
traits measured by a block are varied pairwise, traits not in the pair are varied at fix.levels,
traits not measured by a block are fixed at fix.level.others
}
