% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_info_trace.R
\name{calc.info.trace}
\alias{calc.info.trace}
\title{Trace of information matrix}
\usage{
calc.info.trace(infos, avg = F)
}
\arguments{
\item{infos}{block information as returned from calc.info.block()}

\item{avg}{logical, average result across persons? defaults to FALSE}
}
\value{
matrix of traces of information matrices, rows=persons, columns=blocks
}
\description{
calculates the traits (= sum of diagonal elements) of the information matrix
}
\details{
This is the criterion of A-optimality used in computerized adaptive testing (CAT).
It can be used to summarize information for a block in one number, weighting all precision for all traits equally.
It is inversely proportional to standard errors.
}
