% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pattern_info.R
\name{calc.pattern.info}
\alias{calc.pattern.info}
\title{information for a single ranking pattern}
\usage{
calc.pattern.info(FUN = FUN, traits, ...)
}
\arguments{
\item{FUN}{response probability function}

\item{traits}{vector of traits for a single person}

\item{...}{arguments passed to FUN}
}
\value{
matrix of ntraits x ntraits, information for this pattern
}
\description{
information for a single ranking pattern
}
