% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rel_emp.R
\name{calc.rel.emp}
\alias{calc.rel.emp}
\title{empirical reliability based on factor scores}
\usage{
calc.rel.emp(scores, ses)
}
\arguments{
\item{scores}{matrix of factor scores, rows = persons, columns = traits}

\item{ses}{matrix of SEs, same as scores}
}
\value{
vector of empirical reliabilities for each trait
}
\description{
empirical reliability based on factor scores
}
