% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_testinfo_1d.R
\name{calc.testinfo.1d}
\alias{calc.testinfo.1d}
\title{Testinfo 1D}
\usage{
calc.testinfo.1d(tr.levels, sigma, n = 20, seed = NULL, which.blocks, ...)
}
\arguments{
\item{tr.levels}{vector of trait levels to vary on x-axis}

\item{sigma}{matrix of trait correlations}

\item{n}{numeric, random samples to draw for other traits}

\item{seed}{numeric, optional, random seed}

\item{which.blocks}{vector, blocks for which to calculate testinfo}

\item{...}{FUN for block info and other arguments passed to FUN}
}
\value{
list of 1) matrix of standard errors ("ses") and 2) vector of varied levels ("tr.levels")
}
\description{
calculate testinfo for one trait varied, levels for other traits are randomly sampled
from multivariate normal distribution with covariance sigma, averaged across sample of size n
}
