% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_testinfo_plot.R
\name{calc.testinfo.plot}
\alias{calc.testinfo.plot}
\title{compute testinformation for plots}
\usage{
calc.testinfo.plot(
  tr.levels,
  fix.level.others = 0,
  K,
  loads,
  which.blocks,
  ...
)
}
\arguments{
\item{tr.levels}{vector, trait levels that vary}

\item{fix.level.others}{numeric, level for traits fixed for this plot, defaults to 0}

\item{K}{numeric, number of blocks}

\item{loads}{matrix of item loadings}

\item{which.blocks}{vector, subset of blocks to calcuate testinfo for, defaults to 1:K}

\item{...}{FUN to calculate information and further parameters passed to FUN}
}
\value{
information for relevant trait combinations for all blocks, and for which trait pairs (combinations) information is
}
\description{
computes testinformation for all pairwise trait combinations, other traits are fixed to fix.level.others
}
