% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_block_ind.R
\name{create.block.ind}
\alias{create.block.ind}
\title{block.ind}
\usage{
create.block.ind(K, nb)
}
\arguments{
\item{K}{numeric, number of blocks}

\item{nb}{numeric, block size}
}
\value{
matrix, blocks in columns
}
\description{
create indices for items
}
\examples{
create.block.ind(20, 3)

}
