% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blocks_int.R
\name{create.blocks.int}
\alias{create.blocks.int}
\title{Intercept indices}
\usage{
create.blocks.int(K, nb)
}
\arguments{
\item{K}{numeric, number of blocks}

\item{nb}{numeric, block size}
}
\value{
matrix of intercept indices for each block
}
\description{
create indices for intercepts
}
\examples{
create.blocks.int(20,3)
}
