% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_block.R
\name{create.design.block}
\alias{create.design.block}
\title{generate design matrix for one block}
\usage{
create.design.block(nb)
}
\arguments{
\item{nb}{numeric, block size}
}
\value{
matrix: rows = pairs, columns = items
}
\description{
generate design matrix for one block
}
\examples{
create.design.block(3)
}
