% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_mat.R
\name{create.design.mat}
\alias{create.design.mat}
\title{MFC design matrix}
\usage{
create.design.mat(K, db = NULL, nb = NULL)
}
\arguments{
\item{K}{numeric, number of blocks}

\item{db}{matrix, optional. block design as returned from create.design.block(), if NULL: computed from nb}

\item{nb}{numeric, optional. Block size}
}
\value{
design matrix of mfc: rows=pairwise comparisons, columns=items
}
\description{
MFC design matrix
}
