% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create.grid}
\alias{create.grid}
\title{Create grid of traits}
\usage{
create.grid(tr.levels, ntraits)
}
\arguments{
\item{tr.levels}{vector of trait levels}

\item{ntraits}{integer, number of traits}
}
\value{
dataframe with all combinations of trait levels for ntraits
}
\description{
Create grid of traits
}
\examples{
create.grid(seq(-1,1,.5), 3)
}
