% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_perms_int.R
\name{create.perms.int}
\alias{create.perms.int}
\title{pair intercept indices for each permutation}
\usage{
create.perms.int(nb, perms = NULL)
}
\arguments{
\item{nb}{numeric, block size}

\item{perms}{permutations as returned from permute(). computed internally if NULL.}
}
\value{
matrix: rows = intercepts, columns = permutations
}
\description{
pair intercept indices for each permutation
}
\examples{
create.perms.int(4)
}
