% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Tr.R
\name{create.tr}
\alias{create.tr}
\title{internal: create matrix of sequential comparisons}
\source{
SafirsMFC
}
\usage{
create.tr(nb)
}
\arguments{
\item{nb}{block size}
}
\value{
matrix of sequential comparisons (last row is unimportant)
}
\description{
internal: create matrix of sequential comparisons
}
\examples{
create.tr(3)
}
