% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traits_blocks.R
\name{create.traits.blocks}
\alias{create.traits.blocks}
\title{Extract traits from loading matrix}
\usage{
create.traits.blocks(loads, which.blocks, nb)
}
\arguments{
\item{loads}{matrix of loadings}

\item{which.blocks}{vector of block indices to select}

\item{nb}{numeric, block size}
}
\value{
matrix, rows=blocks, columns=traits
}
\description{
Extract traits from loading matrix
}
