% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info2se.R
\name{info2se}
\alias{info2se}
\title{convert information to SEs}
\usage{
info2se(infos, summed = TRUE, var.out = FALSE)
}
\arguments{
\item{infos}{list of arrays with block information as returned from calc.info.block()}

\item{summed}{logical, should the information be summed across blocks? defaults to TRUE}

\item{var.out}{logival, return variances instead of SEs? defaults to FALSE}
}
\value{
if summed across bocks: matrix: rows=persons, columns=traits,
otherwise: list with entries for each person: matrix of SEs: rows = blocks, columns = traits
}
\description{
across persons
}
