% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh_all.R
\name{lh.all}
\alias{lh.all}
\title{round2}
\usage{
lh.all(
  th,
  int,
  loads,
  uni,
  y_b,
  blocks,
  blocks_int,
  perms,
  perms_int,
  lhb.fun = lhb.fun,
  ...
)
}
\arguments{
\item{int}{vector of intercepts}

\item{loads}{matrix of factor loadings}

\item{uni}{matrix of error covariances}

\item{y_b}{rank order}

\item{perms}{permutations}

\item{perms_int}{indices of intercepts (+ or -)}

\item{lhb.fun}{function to compute likelihood}

\item{traits}{vector of traits}
}
\value{
likelihood of all
}
\description{
likelihood of rank order for all items (and one person) under Thurstonian IRT model
}
