% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh.R
\name{lh}
\alias{lh}
\title{likelihood across all questionnaire items}
\usage{
lh(
  traitsj,
  lhb.fun,
  responsesj,
  bi,
  bi_int,
  int,
  loads,
  uni,
  m.prior,
  s.prior,
  est = "MAP",
  ...
)
}
\arguments{
\item{traitsj}{vector of traits for one person}

\item{bi}{matrix of block indices as returned from create.block.ind()}

\item{bi_int}{matrix of intercep indices as returned from create.blocks.int()}

\item{int}{vector of item intercepts}

\item{loads}{matrix of item loadings, rows = items, columns = traits}

\item{uni}{matrix of item uniquenesses}

\item{m.prior}{vector of means used for multivariate normal prior}

\item{s.prior}{covariance matrix used for multivariate normal prior}

\item{...}{other arguments passed to FUN}

\item{FUN}{function to compute response probability}

\item{perms}{permutations}
}
\value{
list of length N, with entries (block,trait,trait)
}
\description{
across items for one person, maximum a posteriori with multivariate normal prior
}
