% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhb_mplus.R
\name{lhb.mplus}
\alias{lhb.mplus}
\title{Likelihood Mplus}
\usage{
lhb.mplus(
  traits,
  int,
  loads,
  uni,
  y_b,
  perms,
  perms_int,
  Tr,
  nb,
  algorithm = mvtnorm::TVPACK(),
  loga = TRUE
)
}
\arguments{
\item{traits}{vector of traits}

\item{int}{vector of intercepts and minus intercepts for block b}

\item{loads}{matrix of factor loadings for block b}

\item{uni}{matrix of error (co)variances for block b}

\item{y_b}{index for rank order in perms for block b}

\item{perms}{permutations (order(rank(), decreasing=F) order(, decreasing=T))}

\item{perms_int}{indices of intercepts (+ or -)}

\item{Tr}{matrix of pairwise comparisons (only consecutive ones, last row is ignored)}

\item{nb}{numeric, block size}

\item{algorithm}{algorithm to use for numerical approximation in mvtnorm::pmvnorm, defaults to TVPACK()}

\item{loga}{logical: return logarithmized result? defaults to TRUE}
}
\value{
likelihood
}
\description{
likelihood of rank order under Thurstonian IRT model, using pair intercepts as Mplus
}
