% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_items.R
\name{name.items}
\alias{name.items}
\title{name.items}
\usage{
name.items(design.load, prefix)
}
\arguments{
\item{design.load}{matrix: rows = items, cols = traits}

\item{prefix}{prefix for item names, before 01,02...10}
}
\value{
item names
}
\description{
create itemnames for a design load matrix
}
