% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out2blocks.R
\name{out2blocks}
\alias{out2blocks}
\title{out2blocks}
\usage{
out2blocks(y, blocks)
}
\arguments{
\item{y}{data.frame/matrix of pairwise outcomes: rows = participants, columns = outcomes}

\item{blocks}{matrix of block indices, created from block.ind}
}
\value{
matrix of block indices: rows = participants, columns = blocks
}
\description{
pairwise outcomes to block indices
}
