% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out2perm.R
\name{out2perm}
\alias{out2perm}
\title{Outcomes to indices for permutations}
\usage{
out2perm(y, blocks, perms, design.block)
}
\arguments{
\item{y}{matrix of outcomes, rows = participants, columns = pairwise outcomes}

\item{blocks}{matrix, created with block.ind}

\item{perms}{permutations}

\item{design.block}{block design}
}
\value{
matrix of permutation indices, rows = participants, columns = blocks
}
\description{
Outcomes to indices for permutations
}
