% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{permute}
\source{
https://www.r-bloggers.com/learning-r-permutations-and-combinations-with-base-r/
}
\usage{
permute(v)
}
\arguments{
\item{v}{vector to permute}
}
\value{
matrix, permuted elements in columns
}
\description{
generate all possible permutations (without replacement)
}
\examples{
permute(1:3)
}
