% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_block.R
\name{plot.block}
\alias{plot.block}
\title{plot block information}
\usage{
\method{plot}{block}(
  which.blocks,
  info,
  K,
  loads,
  trait.names = NULL,
  trait.cov = trait.cov,
  y = "r2",
  col = "skyblue1",
  bty = "u",
  col.grid = "grey45",
  r = 1,
  phi = 30,
  shade = 0.2,
  ltheta = 60,
  ticktype = "detailed",
  zlab = "\\n\\nR2",
  border = NA,
  nticks = 6,
  scale = TRUE,
  ...
)
}
\arguments{
\item{which.blocks}{vector, indices for blocks to plot}

\item{info}{information for all blocks as returned from calc.info.block()}

\item{K}{integer, number of blocks}

\item{loads}{matrix of item loadings}

\item{trait.names}{names for traits, if NULL defaults to Trait 1, Trait 2 etc.}

\item{...}{further parameters passed to plot3D::persp3D}
}
\value{
plot with reduction in SEs, traits in rows, fixed trait in columns
}
\description{
plot block information
}
