% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_testinfo.R
\name{plot.testinfo}
\alias{plot.testinfo}
\title{plot testinformation}
\usage{
\method{plot}{testinfo}(
  info,
  loads = NULL,
  trait.names = NULL,
  par.mfrow = c(3, 4),
  col = "skyblue1",
  bty = "u",
  col.grid = "grey45",
  r = 1,
  phi = 30,
  shade = 0.2,
  ltheta = 60,
  ticktype = "detailed",
  xlab = "\\nTrait 1",
  ylab = "\\nTrait 2",
  zlab = "\\n\\nSE",
  border = NA,
  nticks = 6,
  scale = TRUE,
  ...
)
}
\arguments{
\item{info}{information as returned from testcalc.info.plot()}

\item{loads}{matrix of item loadings}

\item{trait.names}{vector of trait names, defaults to "Trait 1","Trait 2", ...}

\item{par.mfrow}{mfrow argument for par(), vector with c(rows,columns) for plot}

\item{...}{further parameters passed to persp3D}
}
\value{
plot with reduction in SEs, traits in rows, fixed trait in columns
}
\description{
plots testinformation for all pairwise trait combinations
}
