% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_traits.R
\name{sample.traits}
\alias{sample.traits}
\title{sample traits from multivariate normal distribution with one trait fixed}
\usage{
sample.traits(tr.levels, fix.dim, sigma, n = 10)
}
\arguments{
\item{tr.levels}{vector: levels for fixed trait}

\item{fix.dim}{scalar: dimension to fix}

\item{sigma}{matrix: trait correlations}

\item{n}{scalar: number of samples for each trait level, defaults to 10}
}
\value{
traits correlating as specified in sigma, with dimension fix.dim fixed to tr.levels
}
\description{
sample traits from multivariate normal distribution with one trait fixed
}
\examples{
sample.traits(seq(-2,2,1), fix.dim=2, sigma=diag(3), n=100)
}
