% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_greedy.R
\name{select.greedy}
\alias{select.greedy}
\title{select blocks based on greedy algorithm}
\usage{
select.greedy(
  infos,
  FUN,
  traits.grid,
  K,
  K.start,
  K.final,
  maximize = F,
  weights.grid = NULL,
  constraint.list = NULL,
  ...
)
}
\arguments{
\item{infos}{block information as returned from calc.block.info()}

\item{FUN}{function to summarize block information}

\item{traits.grid}{matrix, traits info was computed for}

\item{K}{integer, number of blocks in pool}

\item{K.start}{integer, number of blocks in initial questionnaires. It is assumed that these are the first K.start blocks.}

\item{K.final}{integer, number of blocks in final questionnaire}
}
\value{
numeric vector of indices of selected blocks
}
\description{
select blocks based on greedy algorithm
}
