% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_optimal.R
\name{select.optimal}
\alias{select.optimal}
\title{Select Optimal Block Combination}
\usage{
select.optimal(
  info.sum,
  traits.grid,
  info.start = NULL,
  K,
  K.final,
  weights.grid = NULL,
  constraint.list = NULL
)
}
\arguments{
\item{info.sum}{summary of block information as returned from calc.info.block}

\item{traits.grid}{grid of traits for which block information was calculated}

\item{K}{integer number of blocks in the pool}

\item{K.final}{integer, final number of blocks}
}
\value{
list: solved should equal 0 if a solution was found, ind.opt contains the indices of selected blocks
}
\description{
Select Optimal Block Combination
}
