% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_responses.R
\name{sim.responses}
\alias{sim.responses}
\title{simulate responses under Thurstonian IRT model}
\usage{
sim.responses(traits, items, design.load, K, nb, return.index = TRUE)
}
\arguments{
\item{traits}{matrix of trait levels, rows = persons, columns = traits}

\item{items}{matrix of item intercepts, loadings, and uniquenesses, as from sim.items()}

\item{design.load}{design matrix of questionnaire, rows=items, columns=traits, 1=positively keyed, -1=negatively keyed}

\item{K}{scalar, number of blocks}

\item{nb}{scalar, block size}

\item{return.index}{logical, return index for rank order as in permute() (TRUE) or rank orders (FALSE)? defaults to TRUE}
}
\value{
matrix of ranks, rows = persons, columns = items
}
\description{
simulate responses under Thurstonian IRT model
}
