% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni2mat.R
\name{uni2mat}
\alias{uni2mat}
\title{Uniqueness matrix}
\usage{
uni2mat(mplus.pars, item.identifier)
}
\arguments{
\item{mplus.pars}{Mplus parameters, read in with MplusAutomation}

\item{item.identifier}{regular expression that is part of all item names}
}
\value{
matrix of estimated uniquenesses (diagonal)
}
\description{
Uniqueness matrix
}
