% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_a_optimality.R
\name{calc.a.optimality}
\alias{calc.a.optimality}
\title{calculate A-optimality}
\usage{
calc.a.optimality(infos)
}
\arguments{
\item{infos}{block information as returned from calc.block.infos}
}
\value{
numeric vector of A-optimality for each person
}
\description{
calculate A-optimality
}
