% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_d_optimality.R
\name{calc.d.optimality}
\alias{calc.d.optimality}
\title{Calculate D-optimality}
\usage{
calc.d.optimality(infos)
}
\arguments{
\item{infos}{block information as returned from calc.info.block()}
}
\value{
numeric vector of D-optimality for each person
}
\description{
Calculate D-optimality
}
